Feature/frontend updates

+++ lib/bookify_web/live/author_live/form_component.ex
@@ -1,5 +1,8 @@
 defmodule BookifyWeb.AuthorLive.FormComponent do
   use BookifyWeb, :live_component
+  import Bookify.Utils.Image
+  import Bookify.Utils.User
+
   alias Bookify.Authors
   alias BookifyWeb.Modules.LiveUploader
 
@@ -29,13 +32,17 @@ defmodule BookifyWeb.AuthorLive.FormComponent do
           <div class="mb-2 text-sm font-semibold">Author Image</div>
           <.live_file_input upload={@uploads.author_image} />
         </div>
-        <%= for entry <- @uploads.author_image.entries do %>
-          <.live_img_preview entry={entry} class="w-48" />
-        <% end %>
-        <%= for {_ref, msg} <- @uploads.author_image.errors do %>
-          <div class="text-red-500 text-lg font-bold">
-            <%= Phoenix.Naming.humanize(msg) <> "!" %>
-          </div>
+        <%= if @uploads.author_image.entries == [] do %>
+          <img src={image(@author)} class="w-48" />
+        <% else %>
+          <%= for entry <- @uploads.author_image.entries do %>
+            <.live_img_preview entry={entry} class="w-48" />
+          <% end %>
+          <%= for {_ref, msg} <- @uploads.author_image.errors do %>
+            <div class="text-red-500 text-lg font-bold">
+              <%= Phoenix.Naming.humanize(msg) <> "!" %>
+            </div>
+          <% end %>
         <% end %>
         <.input field={@form[:wikipedia_url]} type="text" label="Wikipedia url" />
         <:actions>
@@ -70,15 +77,24 @@ defmodule BookifyWeb.AuthorLive.FormComponent do
   end
 
   def handle_event("save", %{"author" => author_params}, socket) do
-    author_params = handle_upload(author_params, socket)
-    save_author(socket, socket.assigns.action, author_params)
+    if current_user(socket) |> is_admin?() do
+      author_params = handle_upload(author_params, socket)
+      save_author(socket, socket.assigns.action, author_params)
+    else
+      not_allowed(socket)
+    end
   end
 
   defp save_author(socket, :edit, author_params) do
+    upload_entries = socket.assigns.uploads.author_image.entries
+
     case Authors.update_author(socket.assigns.author, author_params) do
       {:ok, author} ->
-        LiveUploader.delete_file(socket.assigns.old_filename)
-        notify_parent({:saved, author})
+        if upload_entries != [] do
+          LiveUploader.delete_file(socket.assigns.old_filename)
+        end
+
+        notify_parent({:author_saved, author})
 
         {:noreply,
          socket
@@ -93,7 +109,7 @@ defmodule BookifyWeb.AuthorLive.FormComponent do
   defp save_author(socket, :new, author_params) do
     case Authors.create_author(author_params) do
       {:ok, author} ->
-        notify_parent({:saved, author})
+        notify_parent({:author_saved, author})
 
         {:noreply,
          socket
@@ -115,4 +131,11 @@ defmodule BookifyWeb.AuthorLive.FormComponent do
   end
 
   defp notify_parent(msg), do: send(self(), msg)
+
+  defp not_allowed(socket) do
+    {:noreply,
+     socket
+     |> put_flash(:error, "Not allowed!")
+     |> push_navigate(to: ~p"/")}
+  end
 end
